requirements.txt

discord.py>=2.3.0
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
psutil>=5.9.0
aiohttp>=3.8.0